- renamed a file and corrected paths
