+++ client/src/components/contacts/ContactForm.js
@@ -1,6 +1,6 @@
 import React, { useState, useContext } from "react";
 import ContactContext from "./../context/contact/contactContext";
-import {ContactLabels} from "./ContactLabels";
+import {ContactLabels} from "./Labels";
 
 const ContactForm = () => {
   const contactContext = useContext(ContactContext);
